added authorization for hotel and user routes

+++ src/controllers/hotel.js
@@ -1,17 +1,38 @@
 const Hotel = require('../models/hotel')
 
 const createHotel = async (req, res, next) => {
-    const newHotel = new Hotel(req.body)
+    const user = req.user
     try {
+        if (!user.isAdmin) {
+            throw new Error('you are not admin, please register as admin to perform the operation!')
+        }
+        const newHotel = new Hotel(req.body)
         await newHotel.save()
         res.status(201).send(newHotel)
+
     } catch (e) {
         next(e)
     }
 }
 
+// const createHotel = async (req, res, next) => {
+
+//     const newHotel = new Hotel(req.body)
+//     try {
+//         await newHotel.save()
+//         res.status(201).send(newHotel)
+//     }
+//     catch (e) {
+//         next(e)
+//     }
+// }
+
 const updateHotel = async (req, res, next) => {
+    const user = req.user
     try {
+        if (!user.isAdmin) {
+            throw new Error('you are not admin!')
+        }
         const updateHotel = await Hotel.findByIdAndUpdate(req.params.id, { $set: req.body }, { new: true })
         res.status(200).send(updateHotel)
     } catch (e) {
@@ -20,7 +41,11 @@ const updateHotel = async (req, res, next) => {
 }
 
 const deleteHotel = async (req, res, next) => {
+    const user = req.user
     try {
+        if (!user.isAdmin) {
+            throw new Error('you are not admin!')
+        }
         await Hotel.findByIdAndDelete(req.params.id)
         res.status(200).send('The Hotel has been deleted!')
     } catch (e) {
